requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.11.0
deprecation==2.1.0
Flask==3.0.3
functions-framework==3.8.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mistralai==0.4.1
orjson==3.10.5
packaging==24.1
pydantic==2.7.4
pydantic_core==2.18.4
pyTelegramBotAPI==4.20.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.2
watchdog==4.0.1
Werkzeug==3.0.3
